Refactored node cleanup to seperate file.
